requirements/dev.txt
alembic==1.13.2
crypto==1.4.1
fastapi==0.111.0
passlib==1.7.4
psycopg2==2.9.9
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
starlette==0.37.2
uvicorn==0.30.1